removed pass-in collections for the cluster client
It was removed because of conflicts, for example if shard 0 and shard 1 both shared user x, and shard 0 owns it, if shard 0 got a ready payload, it would reset the users that used shard 0, meaning shard 1 will be using a user that was deleted.

reformatted some stuff

voice connection now creates an opus decoder for each user, the reason why is because if they share one opus decoder, it'll overload and return staticy audio.