0.3.0, breaking changes to `onBefore` and `onCancel`
added `onBeforeRun` and `onCancelRun` which will replace `onBefore` and `onCancel`.

new `onBefore` and `onCancel` is ran before args are parsed for the command, so if you need to check for specific perms in a channel or something, you do it here, then `onBeforeRun` and `onCancelRun` will have the parsed args.